Refactor the logic that removes global ignores
